sprintfix:修改接口部分

+++ apps/good_purchase/views.py
@@ -16,6 +16,7 @@ import uuid
 from datetime import datetime
 from tempfile import NamedTemporaryFile
 
+from apps.good_apply.models import Organization, OrganizationMember, Secretary
 from apps.good_purchase.models import (Cart, Good, GoodType, GroupApply,
                                        UserInfo, Withdraw, WithdrawReason)
 from apps.good_purchase.serializers import (CartSerializer,
@@ -84,10 +85,14 @@ def upload_img(request):
 
 
 class WithdrawReasonViewSet(viewsets.ModelViewSet):
+
     queryset = WithdrawReason.objects.all()
     serializer_class = WithdrawReasonSerializer
 
     def list(self, request, *args, **kwargs):
+        req = request.data
+        org_id = req.get("org_id")
+        self.queryset = WithdrawReason.objects.filter(org_id=org_id)
         withdraw_reason_list = [reason.to_json() for reason in self.queryset]
         return JsonResponse(success_code(withdraw_reason_list))
 
@@ -101,6 +106,7 @@ class WithdrawViewSet(viewsets.ModelViewSet):
         """提交退回物资申请"""
         req = request.data
         username = request.user.username
+        org_id = req.get("org_id")
         check_withdraws_seralizers = CheckWithdrawsSeralizers(data=req)
         if not check_withdraws_seralizers.is_valid():
             message = get_error_message(check_withdraws_seralizers)
@@ -117,20 +123,21 @@ class WithdrawViewSet(viewsets.ModelViewSet):
         else:
             position = '无'
         remark = check_withdraws_seralizers.validated_data.get("remark", '')
-        good_ids = GroupApply.objects.filter(id__in=ids, status=2, username=username).values_list("id", flat=True)
+        good_ids = GroupApply.objects.filter(Q(id__in=ids) & Q(status=2) & Q(
+            username=username) & Q(org_id=org_id)).values_list("id", flat=True)
         if not set(ids).issubset(good_ids):
             raise BusinessException(StatusEnums.SENDBACK_ERROR)
-        if not WithdrawReason.objects.filter(id=reason_id).exists():
+        if not WithdrawReason.objects.filter(Q(id=reason_id) & Q(org_id=org_id)).exists():
             raise BusinessException(StatusEnums.DRAWBACK_ERROR)
         # 创建需要批量添加的withdraw数组
         withdraw_list = []
         for good_id in good_ids:
             withdraw = Withdraw(good_apply_id=good_id, username=username, reason_id=reason_id,
-                                position=position, remark=remark)
+                                position=position, remark=remark, org_id=org_id)
             withdraw_list.append(withdraw)
         with transaction.atomic():
             # 修改个人物资状态
-            GroupApply.objects.filter(id__in=list(good_ids)).update(status=1)
+            GroupApply.objects.filter(Q(id__in=list(good_ids)) & Q(org_id=org_id)).update(status=1)
             # 批量添加物资退回表
             Withdraw.objects.bulk_create(withdraw_list)
         return get_result({"message": "退回商品申请提交成功"})
@@ -196,7 +203,9 @@ class CartViewSet(viewsets.ModelViewSet):
         获取购物车信息
         """
         username = request.user
-        good_list = Cart.objects.filter(username=username)
+        req = request.data
+        org_id = req.get("org_id")
+        good_list = Cart.objects.filter(Q(username=username) & Q(org_id=org_id))
         if not good_list.exists():
             raise BusinessException(StatusEnums.BLANK_ERROR)
         cart_list = get_cart_result(good_list, source="shop")
@@ -209,14 +218,15 @@ class CartViewSet(viewsets.ModelViewSet):
         """
         username = request.user
         req = request.data
+        org_id = req.get("org_id")
         cart_id_list = req.get("cartIdList")
         if not isinstance(cart_id_list, list):
-            cart_all_id = Cart.objects.filter(username=username).values_list("id", flat=True)
+            cart_all_id = Cart.objects.filter(Q(username=username) & Q(org_id=org_id)).values_list("id", flat=True)
             if not set(cart_id_list).issubset(set(cart_all_id)):
                 raise BusinessException(StatusEnums.CART_ADX_ERROR)
-        if not Cart.objects.filter(id__in=cart_id_list, username=username).exists():
+        if not Cart.objects.filter(Q(id__in=cart_id_list) & Q(username=username) & Q(org_id=org_id)).exists():
             raise BusinessException(StatusEnums.CART_NULL_ERROR)
-        Cart.objects.filter(id__in=cart_id_list).delete()
+        Cart.objects.filter(Q(id__in=cart_id_list) & Q(org_id=org_id)).delete()
         return get_result({"code": 200, "message": "删除成功"})
 
     @action(methods=['POST'], detail=False)
@@ -225,6 +235,7 @@ class CartViewSet(viewsets.ModelViewSet):
         更新购物车物资数量信息
         """
         req = request.data
+        org_id = req.get("org_id")
         update_goods_list = req.get("goodsList")
         if isinstance(update_goods_list, list):
             update_goods_ids = []
@@ -235,7 +246,7 @@ class CartViewSet(viewsets.ModelViewSet):
                 raise BusinessException(StatusEnums.CART_UPDATE_ERROR)
             all_goods = []
             for good in update_goods_list:
-                temp_good = Cart.objects.get(id=good["id"])
+                temp_good = Cart.objects.get(id=good["id"], org_id=org_id)
                 temp_good.num = int(good["num"])
                 temp_good.update_time = datetime.now()
                 all_goods.append(temp_good)
@@ -252,15 +263,17 @@ class CartViewSet(viewsets.ModelViewSet):
         req = request.data
         good_info = req.get("goodInfo")
         username = request.user
+        org_id = req.get("org_id")
         if isinstance(good_info['num'], int) and good_info['num'] > 0:
-            temp_good = Cart.objects.filter(good_id=good_info["id"], username=username)
+            temp_good = Cart.objects.filter(Q(good_id=good_info["id"]) & Q(username=username) & Q(org_id=org_id))
         else:
             raise BusinessException(StatusEnums.CART_NUM_ERROR)
         if temp_good.exists():
             num = int(temp_good[0].num) + int(good_info["num"])
-            Cart.objects.filter(good_id=good_info["id"]).update(num=num, update_time=datetime.now())
+            Cart.objects.filter(Q(good_id=good_info["id"]) & Q(org_id=org_id)
+                                ).update(num=num, update_time=datetime.now())
         else:
-            Cart.objects.create(good_id=good_info['id'], username=username, num=good_info['num'])
+            Cart.objects.create(good_id=good_info['id'], username=username, num=good_info['num'], org_id=org_id)
         return get_result({"code": 200, "message": "物资成功加入购物车"})
 
 
@@ -273,7 +286,9 @@ class GroupApplyViewSet(viewsets.ModelViewSet):
         """
         获取组内物资信息
         """
-        apply_list = GroupApply.objects.filter(status=4)
+        req = request.data
+        org_id = req.get("oeg_id")
+        apply_list = GroupApply.objects.filter(Q(status=4) & Q(org_id=org_id))
         if not apply_list.exists():
             return get_result({"code": 200, "data": [], "message": "申请状态为购买中的列表为空"})
         cart_list = get_cart_result(apply_list, source="apply")
@@ -286,10 +301,11 @@ class GroupApplyViewSet(viewsets.ModelViewSet):
         """
         req = request.data
         apply_id = req.get("applyId")
+        org_id = req.get("org_id")
         if not check_param_id(apply_id):
             raise BusinessException(StatusEnums.CART_ADX_ERROR)
         try:
-            GroupApply.objects.get(id=apply_id).delete()
+            GroupApply.objects.get(id=apply_id, org_id=org_id).delete()
         except GroupApply.DoesNotExist:
             raise BusinessException(StatusEnums.CART_DELETE_ERROR)
         return get_result({"code": 200, "message": "删除成功"})
@@ -302,6 +318,7 @@ class GroupApplyViewSet(viewsets.ModelViewSet):
         req = request.data
         apply_list = req.get("applyList")
         update_type = req.get("updateType")
+        org_id = req.get("org_id")
         if isinstance(apply_list, list) and check_apply_update_param(update_type):
             update_apply_ids = []
             for update_apply in update_apply_ids:
@@ -311,7 +328,7 @@ class GroupApplyViewSet(viewsets.ModelViewSet):
                 raise BusinessException(StatusEnums.CART_UPDATE_ERROR2)
             all_applies = []
             for apply in apply_list:
-                temp_apply = GroupApply.objects.get(id=apply["id"])
+                temp_apply = GroupApply.objects.get(id=apply["id"], org_id=org_id)
                 temp_apply.num = int(apply["num"])
                 if update_type == 'status':
                     temp_apply.status = 5
@@ -337,7 +354,7 @@ class GroupApplyViewSet(viewsets.ModelViewSet):
         page_limit = int(request.GET.get('pageLimit', 10))
         page = int(request.GET.get('page', 1))
         id_list = request.GET.get('idList', None)
-
+        org_id = req.get("org_id")
         personal_serializer = personalSerializer(data={
             "username": username
         })
@@ -345,7 +362,7 @@ class GroupApplyViewSet(viewsets.ModelViewSet):
             raise ValueError(get_error_message(personal_serializer))
 
         # 获得查询集
-        queryset = GroupApply.objects.filter(username=username)
+        queryset = GroupApply.objects.filter(Q(username=username) & Q(org_id=org_id))
 
         unnecessary_goods = []  # 用于记录被过滤掉的物品
         # 获取form的内容
@@ -381,10 +398,10 @@ class GroupApplyViewSet(viewsets.ModelViewSet):
         if location and location != '0':
             query = query & Q(position__icontains=location)
         if status and int(status) != 0:
-            query = query & Q(status=status)
+            query = query & Q(status=status) & Q(org_id=org_id)
         queryset = queryset.filter(query)
         if good_type and int(good_type) != 0:
-            goods = Good.objects.filter(good_type_id=good_type, status=1)
+            goods = Good.objects.filter(Q(good_type_id=good_type,) & Q(status=1) & Q(org_id=org_id))
             good_codes = []
             for good in goods:
                 good_codes.append(good.good_code)
@@ -427,11 +444,12 @@ class GroupApplyViewSet(viewsets.ModelViewSet):
         "确认收货"
         body = request.data
         id_list = body.get('idList')
+        org_id = body.get("org_id")
         serializer = ConfirmReceiptSerializer(data={"id_list": id_list})
         if not serializer.is_valid():  # 校验参数
             err_msg = get_error_message(serializer)
             return get_result({'result': False, 'message': err_msg})
-        queryset = GroupApply.objects.filter(id__in=id_list, status=5)  # 获取查询集
+        queryset = GroupApply.objects.filter(Q(id__in=id_list) & Q(status=5) & Q(org_id=org_id))  # 获取查询集
         if len(queryset) != len(id_list):  # 列表id中存在状态为非待收货的物品
             raise BusinessException(StatusEnums.PARAMS_ERROR)
         with transaction.atomic():
@@ -450,11 +468,13 @@ class GoodViewSet(viewsets.ModelViewSet):
           """
         req_data = request.GET
         good_id = req_data.get("good_id", 0)
+        org_id = req_data.get("org_id")
+
         # 校验参数
         if not check_param_id(good_id):
             raise BusinessException(StatusEnums. GOODID_ERROR)
         try:
-            good = Good.objects.get(id=good_id, status=1).to_json()
+            good = Good.objects.get(id=good_id, status=1, org_id=org_id).to_json()
         except Good.DoesNotExist:
             raise BusinessException(StatusEnums.NOTFOUND_ERROR)
         return JsonResponse(success_code(good))
@@ -466,7 +486,9 @@ class GoodViewSet(viewsets.ModelViewSet):
         """
         req_data = request.GET
         # 筛选项参数拼接查询条件
-        query = Q(status=1)
+        # username = request.user.username
+        org_id = req_data.get("org_id")
+        query = Q(status=1) & Q(org_id=org_id)
         # 需要筛选的字段
         conditions = ('good_code', 'good_name')
         for condition in conditions:
@@ -502,17 +524,18 @@ class GoodViewSet(viewsets.ModelViewSet):
             raise BusinessException(StatusEnums.AUTHORITY_ERROR)
         good = request.data
         # 参数校验
+        org_id = good.get("org_id")
         good_serializers = GoodSerializers(data=good)
         if not good_serializers.is_valid():
             message = get_error_message(good_serializers)
             return get_result({"code": 7100, "result": False, "message": message})  # 输出错误方式
         # 检查商品编码是否存在
         good_code = good_serializers.validated_data.get("good_code")
-        if Good.objects.filter(good_code=good_code, status=1).exists():
+        if Good.objects.filter(Q(good_code=good_code) & Q(status=1) & Q(org_id=org_id)).exists():
             raise BusinessException(StatusEnums.INPUT_GOODS_ERROR)
         # 检验商品类型是否存在
         good_type_id = good_serializers.validated_data.get("good_type_id")
-        if not GoodType.objects.filter(id=good_type_id).exists():
+        if not GoodType.objects.filter(Q(id=good_type_id) & Q(org_id=org_id)).exists():
             raise BusinessException(StatusEnums.CARTTYPE_NULL_ERROR)
         with transaction.atomic():
             Good.objects.create(**good)
@@ -525,28 +548,29 @@ class GoodViewSet(viewsets.ModelViewSet):
         if not flag:
             raise BusinessException(StatusEnums.AUTHORITY_ERROR)
         good = request.data
+        org_id = good.get("org_id")
+        good_id = good.get("id", 0)
         # 参数校验
         good_serializers = GoodSerializers(data=good)
         if not good_serializers.is_valid():
             message = get_error_message(good_serializers)
             return get_result({"code": 1, "result": False, "message": message})
         # 校验id
-        good_id = good.get("id", 0)
         if not check_param_id(good_id):
             raise BusinessException(StatusEnums.GOOD_ID_ERROR)
         # 校验商品是否存在
-        if not Good.objects.filter(id=good_id, status=1).exists():
+        if not Good.objects.filter(Q(id=good_id) & Q(status=1) & Q(org_id=org_id)).exists():
             raise BusinessException(StatusEnums.CART_NO_ERROR)
         # 检验商品编码是否重复(good_code相同，但id不同的商品)
         good_code = good_serializers.validated_data.get("good_code")
-        if Good.objects.filter(Q(good_code=good_code) & ~Q(id=good_id) & Q(status=1)).exists():
+        if Good.objects.filter(Q(good_code=good_code) & ~Q(id=good_id) & Q(status=1) & Q(org_id=org_id)).exists():
             raise BusinessException(StatusEnums.INPUT_GOODS_ERROR)
         # 校验商品类型是否存在
         good_type_id = good_serializers.validated_data.get("good_type_id")
-        if not GoodType.objects.filter(id=good_type_id).exists():
+        if not GoodType.objects.filter(Q(id=good_type_id) & Q(org_id=org_id)).exists():
             raise BusinessException(StatusEnums.CARTTYPE_NULL_ERROR)
         # 更新数据
-        Good.objects.filter(id=good_id).update(**good, update_time=datetime.now())
+        Good.objects.filter(Q(id=good_id) & Q(org_id=org_id)).update(**good, update_time=datetime.now())
         return get_result({"message": "修改商品信息成功"})
 
     @action(methods=['post'], detail=False)
@@ -557,19 +581,22 @@ class GoodViewSet(viewsets.ModelViewSet):
             raise BusinessException(StatusEnums.AUTHORITY_ERROR)
         req = request.data
         good_id = req.get('id', 0)
+        org_id = req.get("org_id")
         # 校验参数
         if not check_param_id(good_id):
             raise BusinessException(StatusEnums.GOODID_ERROR)
         # 校验商品是否存在
-        if not Good.objects.filter(id=good_id, status=1).exists():
+        if not Good.objects.filter(id=good_id, status=1, org_id=org_id).exists():
             raise BusinessException(StatusEnums.CART_NO_ERROR)
-        Good.objects.filter(id=good_id).update(status=0)
+        Good.objects.filter(Q(id=good_id) & Q(org_id=org_id)).update(status=0)
         return get_result({"message": "下架商品成功"})
 
     @action(methods=['get'], detail=False)
     def get_good_code_list(self, request):
         """获取商品编码列表"""
-        good_list = Good.objects.all()
+        req = request.data
+        org_id = req.get("org_id")
+        good_list = Good.objects.filter(org_id=org_id)
         good_type_list = [good_item.good_code for good_item in good_list]
         return JsonResponse(success_code(good_type_list))
 
@@ -585,21 +612,53 @@ class GoodTypeViewSet(viewsets.ModelViewSet):
         if not flag:
             raise BusinessException(StatusEnums.AUTHORITY_ERROR)
         req = request.data
+        org_id = req.get("org_id")
         # 参数校验
         good_type_serializers = GoodTypeSerializers(data=req)
         if not good_type_serializers.is_valid():
             message = get_error_message(good_type_serializers)
             return get_result({"code": 7100, "result": False, "message": message})
-        # 验证该商品类型是否存在
+        # 验证该商品类型在该组是否存在
         type_name = good_type_serializers.validated_data.get("type_name")
-        if GoodType.objects.filter(type_name=type_name).exists():
+        query = Q(type_name=type_name) & Q(org_id=org_id)
+        if GoodType.objects.filter(query).exists():
             raise BusinessException(StatusEnums.INPUT_TYPE_ERROR)
-        good_type = GoodType.objects.create(type_name=type_name)
-        return get_result({"message": "新增商品类型成功", "data": {"id": good_type.id}})
+        good_type = GoodType.objects.create(type_name=type_name, org_id=org_id)
+        return get_result({"message": "新增商品类型成功", "data": {"id": good_type.id, "org_id": org_id}})
 
     @action(methods=['get'], detail=False)
     def get_good_type_list(self, request):
         """获取商品类别列表"""
-        good_types = GoodType.objects.all()
+        req = request.data
+        org_id = req.get("id")
+        good_types = GoodType.objects.filter(org_id=org_id)
         good_type_list = [good_type.to_json() for good_type in good_types]
         return JsonResponse(success_code(good_type_list))
+
+
+class OraganizationViewSet(viewsets.ModelViewSet):
+    @action(methods=['post'], detail=False)
+    def create_group(self, request):
+        """创建"""
+        username = request.username
+        req = request.data
+        group_name = req.get("group_name")
+        org = Organization.objects.create(group_name=group_name)
+        Secretary.objects.create(username=username, org_id=org.id)
+        OrganizationMember.objects(username=username, org_id=org.id)
+        return get_result({"message": "创建组成功"})
+
+    @action(methods=['get'], detail=False)
+    def get_all_user_groupid(self, request):
+        """获取用户所有所在组接口"""
+        username = request.username
+        """根据用户名获取他所在的组成员表查询出来"""
+        group_user = OrganizationMember.objects.filter(username=username)
+        """循环得到"""
+        group_user_id_list = [gro.to_json() for gro in group_user]
+        # def to_json(self):
+        #     return {
+        #         "id": self.id,
+        #         "type": self.type_name
+        #     }
+        return JsonResponse(success_code(group_user_id_list))
